(PIE-799) Changes to allow script to run on agent
